move sa sync code to existing sync.go file to be all together
